requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.19.0
pydantic>=1.10.0,<2.0.0
PyYAML>=6.0
rapidfuzz>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
aiofiles>=23.0.0
python-multipart>=0.0.6
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
